Plugin: move some stuff to dsl package
