+++ rxviper-gradle-plugin/src/main/kotlin/com/dzaitsev/rxviper/plugin/internal/dsl/ScreenFactory.kt
@@ -1,5 +1,6 @@
-package com.dzaitsev.rxviper.plugin
+package com.dzaitsev.rxviper.plugin.internal.dsl
 
+import com.dzaitsev.rxviper.plugin.aClass
 import org.gradle.api.NamedDomainObjectFactory
 import org.gradle.api.Project
