+++ rxviper-gradle-plugin/src/main/kotlin/com/dzaitsev/rxviper/plugin/internal/dsl/UseCase.kt
@@ -1,4 +1,6 @@
-package com.dzaitsev.rxviper.plugin
+package com.dzaitsev.rxviper.plugin.internal.dsl
+
+import com.dzaitsev.rxviper.plugin.aClass
 
 class UseCase(_name: String) {
   internal val name = _name.capitalize()
