+++ rxviper-gradle-plugin/src/test/kotlin/com/dzaitsev/rxviper/plugin/UseCaseTest.kt
@@ -1,5 +1,6 @@
 package com.dzaitsev.rxviper.plugin
 
+import com.dzaitsev.rxviper.plugin.internal.dsl.UseCase
 import com.google.common.truth.Truth.assertThat
 import org.junit.Test
